Provide "read-only" mode for management script
If `RELEASE_READ_ONLY` is set, management script will avoid generating
any files, or executing tasks which require files to be generated. Hooks
and custom commands are run unrestricted though, so it is up to the user
to make sure that those are compatible with this mode.

The use case here is described in #360

**IMPORTANT:** This read-only mode is a courtesy, not a guarantee, as we
can only control what Distillery tries to generate, so bear that in
mind.

Closes #360